Added back the Executing directory using the Assembly Code Base.
The main problem encountered before is that when the Executing Assembly cannot be found, we use the CurrentDomain.BaseDirectory, which can be “/“ (root) if the application is executed from root. What we want with the ExecutingDirectory value is the actual location of the Executing Application. When the assembly can be loaded, then we use the Assembly.CodeBase.